Removed all admin roles, added integration specs
There was no need for this engine to even know about admin users, much
less require them. So I have removed that from here.

Added integration specs that exercise the very simple dummy rails app.
That app has a `User` class, but it is just a silly in memory model.