Add proper support for provider url
This required rolling in the omniauth strategy into this gem (and no
longer requiring the omniauth-radius one). I am not sure we want to keep
it this way, but this keeps us moving for now.

Move things into an environment file, where we read the env for
RADIUS_OAUTH_PROVIDER_URL and set it to a constant. That is used in the
strategy as well as the engine.