Refactor Login and Updater to be Authenticator
This creates an Authenticator the functions as a factory of the host
app's User model. This class will provide fetchers that retrieve the
auth_hash from the Kracken server and use that to find or create the
User model. Most of the time these find-or-creaters will return the
model if it finds one, or nil if it does not. In other cases it should
raise an exception.

Tests were cleaned up, and better exercise the api integration cases.
Including some minor maintenance to the dummy rails app.

Added support for BugSnag notifications if it is included.