Automatically update the user if it has changed on Kracken
We needed a way to update the user information on kracken and
automatically update all the client apps. Instead of pushing changes
out to them we added a cookie that will act as an indicator that
the user is stale and they need to be updated.

The refresh is accomplished by redirecting to the normal oauth flow
which will simply redirect the back if they are already signed in (or
ask for a user/pass if they are not).

The basic steps are:

 - Check for the `_radius_user_cache_key` tld cookie
 - Compare it to the `user_cache_key` in the session
 - If they don't match, redirect them to the oauth provider and
   delete the cookie