Handle the user cache cookie on public pages
Any pages that didn't require a current user were ignoring the cache
key cookie. This change will look for the existence of that and update
or log out the user as appropriate.

The logic is a bit murky, but with my local testing it appears to work
in all the edge cases, certainly something that could be cleaned
up.