Improved API Support from downstream
This merges support from changes and improvements made in Sauron to be
part of this library.

This adds middleware the Rails rack stack. It sits after the Rails debug
exceptions middleware so it can intercept errors prior to Rails handling
them. It also seem to move our middleware after the other debug
middleware plugins. This is good for the test environment so we actually
see our JSON error messages.

There is a bit of boiler plate setup code here just to make Rails happy.

This also moves the previous custom resource controller code out into
the lib directory prepping it for extraction.

This adds error pages for beacons and team resources in the API. This
conforms to the JSON API by adding:

- a 409 code if there is a conflict
- always rendering an array of errors
- showing the error attr path in the request

In order to handle the conflict a new module is available to use. It
sits on top of the uniqueness validator. It uses it to shim in what
attrs may or may not have a uniqueness conflict. This is necessary to
check if a specific attribute was a conflict instead of relying on
inspecting the generated message.

No current resources in the public API use the conflict validator.