Raise better exception based on token validation with server
Having the token fail should not raise a request error, so capture a 401
response code and raise that as a specific TokenUnauthorized exception.
Other errors will raise a RequestError (or possibly some other one) but
that should be much more descriptive when troubleshooting errors.

Move `request_http_token_authentication` to the token module, since it
have more to do with tokens than it does with API.