Only cache the user_id when parsing the auth_hash
We don't want stale user models being pulled from the cache. So only
save the `user_id`.

We don't want to query the database twice. So create a local variable
for the user, set it to nil, fetch from cache and only query if there
was a cache-hit (thus user is still nil).