Brain dump cache related behavior
This is the spec outline for how I believe the new caching layer should
behave. From this we can see there is some overlap in behavior across
the states. Additionally, it seems we may have some slight adjustments
to the existing behavior. I believe we can prove that everything is
backwards compatible though, so it should not be an issue.