Setup basic spec state and extract common behavior
This is in preparation for the new caching changes. Since the default
test cache is the [`NullStore`](http://api.rubyonrails.org/classes/ActiveSupport/Cache/NullStore.html)
we need to ensure we test the cache behavior. Since we don't know what
side-effects enabling the cache for the entire test suite will result
in, we enable it only for these specs. Since it's likely that we will
want to test caching in other parts of the code later, we make put this
common setup into a support module.

We choose the in-memory cache as these are tests and we want them to be
fast. The in-memory cache requires no external dependencies and is one
of the fastest cache options.

We know all forms of the cache state will still require munging of the
headers. So since we're extracting common behavior we moved the header
munging logic in to a shared example set.