Handle subsequent cache reads which are frozen
More quirks of the Rails cache. Initial testing showed that writing a
frozen cache value returns an unfrozen value. Likely due to poor rigor
in test methodology previous thought was that all reads of the cache
would result in unfrozen values - even if that cache'd entry was
refrozen.

This turns out to be false. The actual behavior is:

- Write frozen object to cache
- Read cache receiving an unfrozen version
- All subsequent cache reads will be unfrozen until one of the read
  versions is re-frozen
- Subsequent cache reads will be frozen

This was causing "500 Internal Server" errors in what appeared to be
ActiveRecord calls. Digging into the stack trace showed it was our
caching.