Fix support for Ruby 2.1
It seems Ruby changed how `nil` behaves between 2.1 and 2.2.

- Ruby 2.1

  ```ruby
  nil.frozen?
  # => false
  ```
- Ruby 2.2

  ```ruby
  nil.frozen?
  # => true
  ```

We anticipate that most values will not be `nil` and most of the Rails
apps are already on Ruby 2.2. Additionally, under heavier loads all but
the first read will already be frozen. So we place the `nil` check as
the 2nd predicate as we usually won't need it.