Fix mutation of hash.
There was no reason to be mutating the hash. This is what was actually
causing the `RuntimeError: can't modify frozen Hash` problem. What we
simply wanted was to freeze the auth hash and the inner team array. To
protected against future shallow values we just freeze all the inner
values.

Freezing an already frozen object is a no-op and thread safe; it will
not raise an error. With this new logic we don't have to worry about
threading. Though two or more threads may re-freeze the same object, it
will not cause a crash.

We chose to use `each` here, instead of `values.each`, so that we don't
have to create the intermediate value array.