don't store a reference to the User class
The issue: `Kracken::Config` was storing a reference to the `User`
class. In development, if Rails detects that an app's source has
changed, it would reload the source (including the User class).  This
would leave `Kracken` with a reference to an old version of the `User`
class.

The (partial) fix: don't store a reference. When we need the class, get
it via `::User` and don't store it in an instance variable. This wont
solve the case where the app sets a custom `User` class, but none of our
apps do this currently.