Fix :bug: with message format
(>ლ) we did the work to convert the ids to arrays, but then in the error
message, ignored it. The parameter is `missing_ids` and our ivar is
`@missing_ids`, with reader `#missing_ids`. Since the parameter name is
the same as the reader, references to the plain `missing_ids` is the
parameter not the reader (which would be `self.missing_ids` in this
context).

Instead of appending `self` and passing messages, this makes the message
interpolation use the local ivars. This is acceptable to me as we are
still in the constructor. We just set the ivars above and then want to
use them directly for the message interpolation. I view this as treating
the ivars akin to locals (just pretend that `@` is part of the local var
name) for this specific constructor use case.