Move class method and add module access control
The `included` class method was placed between two module definitions.
This made it difficult to find lexically. This just moves it, white
space update, to after all modules are defined and before the module
methods are defined.

This also marks the module's helper methods as `module_function` access.
This means when they get included they will be private, thus not exposed
to potential routing conflicts.