Port the public exceptions injection
Instead of making every app using our gem insert our custom middleware,
we just load it for them. The standard hook to do this is via a railtie.
We load the railtie when the gem is loaded and we detect we are in a
Rails environment. Rails will do this in the standard
`config/application.rb` when it loads all the gems.