Broaden error detail environments
Normally `consider_all_requests_local` is setup for development
type environments. In these environments we want to get as much
information about our problems as we can. By separating this check from
the status code check we can be sure we get the information we need when
developing new routes, actions, views, etc. Interestingly, the test
environment does not consider the requests local by default. We do not
make an exception because it allows tests to verify the actual returned
responses as a client would see them.

`500 Internal Server` errors are conditions we do not expect. If this
happens during a test we want to ensure we have all the information we
need to properly diagnose the issue.