Pass through non-JSON api errors
This fixes a small :bug: where all errors seem to originate from the
gem. This is because we were catching, then re-raising, errors which
came from non-JSON api endpoints; effectively all parts of the
application.

This changes that behavior so that we are just a noop middleware for
non-JSON api endpoints. This means if they raise an error, we won't
catch it, and the error will appear as coming from it's original source.