Remove ApplicationController from SessonsController
Turns out that rails 5 had different autoloading behavior, and was
autoloading `Kracken::ApplicationController` when it referenced
"ApplicationController` in the `Kracken` module. I am not sure who had
the better behavior, rails 4 or rails 5, but this could be resolved by
dropping the dependency on the `ApplicationController` and just inheriting
from `ActionController::Base` directly.

Once I did that just needed to remove the `skip_before_filters` and
things seemed to work.