Add backwards compatibility with Rails 4
Most of the consuming apps are on Rails 4, so it is important that this
gem continue to work with them.

There are two main changed in this commit.

First to fix the monkey patch to ActionController's
`request_http_token_authentication` because it changes airty between
Rails 4 and Rails 5. It adds a second optional param, but rails won't
let you override a method unless the signature matches -- so this
required two different `def`s to set this based on the Rails version
number. I looked for a way to get the airty of a module method but after
20 min gave up and just checked the rails version number instead.

The second is to use the different syntax with ActionDispatch's
integration tests syntax. The `get` method now expects `params` and
`headers` parameters to be named parameters. Again, I just gave up and
checked the Rails version and made the decision. This was needed because
I was juggling the rails versions in the gemspec to test compatibility.