Use `::User` for the `user_class` in the auth module
This removes the ability to set custom user class in kracken's config.
So you will no longer be able to do this:

```
Kracken.config.user_class = Account
```

Instead it just assumes a AR model called `::User`.

This was required because auto-reloading was broken in Rails 5. Changing
the user model or logging in would cause the following exception to be
raised:

```
A copy of User has been removed from the module tree but is still active!
```

Or even more helpful:

```
Cannot assign property of type User to object of type User
```

The module would store a reference the `User` class and that wouldn't be
reinitalized between requests. To work around this, we just use the
constant.

Luckily all our Rails apps follow that convention.