Add token authorization spec helper
This adds a new RSpec authorization helper designed specifically for API
endpoints. This leverages the fact that token authorization is
temporarily cached to avoid making an external API request to the
configured kracken server.

This type of helper is necessary as the newer Rails 5
`ActionController::API` base class does not mix in cookie or session
support by default. Because of this the existing `sign_in` helper is not
properly suited for these controllers.

Additionally, for API endpoints the `sign_in` helper hides the fact that
authentication _must_ be provided through the `Authorization` header;
instead of assuming it was set previously in the session. This could
cause some edge cases bugs to slip through where authorization isn't
getting set, checked, or goes through the proper process.