Clear token auth cache before specs
We don't want to accidentally leak auth state across the specs. However,
we can't be sure what other information was stored in the cache which
should remain. To be a good Ruby citizen we only clear the cache related
to our auth changes.

We could have kept track of what was added by our mixin. However that is
much more error prone and would require logic to track all the changes.
The simple way to handle this is to just clear all caches which match
our namespace.