Fix `current_user` for use with `token_authorize`
The old token auth logic, which mimiced the session based behavior, did
it by overwriting methods on the model. Now that we are using the
`token_authorize` helper to we have an expectation that the remainder of
the `TokenAuthenticatable` module works as designed. Unfortunately it
doesn't.

Ideally the patching of the `TokenAuthenticatable` module would simply
be removed. However, that would break all of the existing specs which
rely on it (as they are not properly sending the auth headers). Until we
fix those we need to continue to support this wonky old style.

The original `TokenAuthenticatable` looks the user up when
`current_user` is called. It does this by calling `find` which raises an
exception when the user can't be found. This would break our specs as
the patched `authenticate_user_with_token!` will end up calling `find`
before appropriate auth data is set; thus raising the exception. To get
around this we check to see if this data is populated (thus assuming
token authentication was successful) before delegating to the original
logic calling `find`.