Drop custom request class
We only use this class when we check if a request is for a JSON API
endpoint. After that it is not used

After upgrading to Rails 5, enough of the internals changed making this
custom class difficult to support. I think initially the idea was this
would be a Rack only mix-in. However, at this point we use enough of
`ActionDispatch` / `ActiveSupport` to make it a bit of work to be pure
Rack. Since we're accepting that we're a Rails plugin let's just embrace
it and use `ActionDispatch` objects directly; thus we drop our custom
class.