Check content types before processing API endpoint
Now that we are only checking the path before wrapping exceptions we
still need to verify that the endpoint supports the JSON API spec. We do
this via mime type negotiation when processing the request. If we do not
have a valid type then we raise the customary Rails `UnknownFormat`
exception.