Conform to Rails 5 exception wrapper API
The Rails 5 exception wrapper now takes a backtrace cleaner instead of
the environment. Additionally, it attempts to inspect the provided
exception to "find the original exception". This poses an issue for us
as we often catch Rails specific errors and re-raise them with our
custom exceptions. When it comes time to wrap the exception we see the
Rails exception instead of our custom exception. To properly handle this
we record the actual exception raised so we can inspect it for
additional behavior.

Not that it matters but the Rails "find the original exception" is
really "find the parent exception". So if an exception is wrapped more
than once it won't actually be the "original exception":

```ruby
A = Class.new(StandardError)
B = Class.new(StandardError)
C = Class.new(StandardError)

begin
  begin
    begin
      raise A
    rescue
      raise B
    end
  rescue => initial
    initial.class
    # => B

    # This is what Rails sees / uses
    initial.cause.class
    # => A
    raise C
  end
rescue => e
  e.class
  # => C

  # This is what Rails sees / uses
  e.cause.class
  # => B

  # This is the actual original exception
  e.cause.cause.class
  # => A
end
```