Fix state cleanup
This is a :hammer: for ensuring the `current_user` state is cleared.
Currently we are missing clearing this on `feature` and the new `system`
tests. We could add more `before(:type)` blocks to clear the state. But
that would also mean we may not clear the state when someone else
manually includes the helper.

As we already have a generic `before(:example)` hook to clear the token
auth, there isn't a lot of additional overhead to use this to also clear
the `current_user` state. This also ensures we clear everything even for
those cases where the module was manually included.