Fix method redefined warning
This fixes the following warning:

  omniauth-radius-rails/lib/kracken/config.rb:14: warning: method redefined; discarding old user_class

It would be good to fix things to support customizing the user class for
long term flexibility, but it was changed due to some issues with Rails
autoloading retaining old class references. Instead we document it is
likely deprecated and that the reader returns the static constant now.