Switch to in-memory caches for auth
The main reason for this switch is security. By keeping the data only
in-memory we do not have to worry about an external system being
compromised. Additionally, since each server will only have some of the
data we limit the scope of what may be compromised (though really, if
the web server is compromised there are a lot more issues).

Another reasons for this switch is we do not control what type of Rails
cache may be used here. The default Rails cache is file system.
Unfortunately, the file system cache is not thread safe and has some
issues with being multi-process safe (regarding incrementing and
multiple key management).  While Redis is thread/process safe, there's
no guarantee that is the cache type used.

Due to the type and size of the data being stored in this cache, 5
megabytes should be more than enough. This would allow caching tens of
thousands of tokens. Since the main goal of these caches is to help
speed up repeated API requests for the same tokens, this size should be
a sufficient balance between what it can hold and memory overhead.