Add task to clean up old OAuth credentials
This is a helper task to clean up old OAuth tokens. There are
potentially two types of tokens:

  - expired tokens
  - legacy tokens (those without an expiration; these are not personal
    tokens)

We are phasing out the legacy tokens for security reasons. In the future
all OAuth tokens will have fixed expiration windows. This simply means
we'll expect OAuth apps to have the user re-authenticate every so often.
Looking at other companies policies:

  - [Facebook uses 90 days](https://developers.facebook.com/docs/facebook-login/access-tokens/refreshing)
  - [Google uses 6 months](https://developers.google.com/identity/protocols/OAuth2#expiration)

At this time we feel 90 days is a good default. The task is configurable
depending on the app's requirements.