use if/elsif/else logic to force use of `==` operators
Since `request.format` is of type `Mime::Type` which implements a `==`
operator so that it is easy to compare to a symbol, we need to use an if
statement so we can ensure that `Mime::Type`'s `==` operator is called
instead of `Symbol`'s.  Also update the spec to use `Mime::Type` objects
so we're not masking what is happening.