Drop flash from sign out action
The flow redirects to the account server. The account server may not be
redirecting back. Thus the flash will only appear the next time the user
hits the service site, which may actually be after they log in - thus
providing a very confusing flash notice. Instead of trying to use
cookies or something else to handle cross origin messages this simply
drops the flash notice so no message will appear.