Add `sign_out` helper for request specs
It's not often this is necessary, but when there is shared state that
needs to be undone it may be necessary. Prior to this the same thing can
be accomplished using `sign_in nil` but that's a bit clunky. This adds
the more intentional revealing `sign_out` helper, which also provides
better interface parity between the spec types.