Extract shared sign-in shim logic
This makes it easier in the future to provide the `sign_out` parity when
we want include the `sign_in` helper. We name the module as a "shim" to
help indicate the behavior of the module; i.e. shimming in some logic,
instead of changing a configuration, intercepting remote auth calls,
etc.