Add radius-spec gem
This adds and configures our radius-spec gem. One difference to the
standard config is we need to tell Rubocop we are targeting an old
version. This is necessary so it doesn't suggest syntax (i.e. safe
navigator operator) which would break on those versions.

Hopefully, we can drop the older versions very soon.