Add basic binstubs
Normally for gems we do not include binstubs to allow developers to
decide for themselves how to setup their environment. However, in order
to add reviewdog we use a `bin/ci-code-review` custom script. Placing
that in `bin` is standard across our repos. It is also standard for us
to have these common binstubs. This includes them to prep for our CI
scripts.

As we do not include any executables, if we did they would go in `exe/`,
there is no conflict to be concerned about. However, just to be safe,
this also updates the gemspec to be explicit about where executables
live.