Silence warnings when running RSpec with warnings
Previously the test run began with something like this:

    kracken/rspec.rb:52: warning: method redefined; discarding old current_user
    kracken/controllers/authenticatable.rb:97: warning: previous definition of current_user was here

> Don't Live with Broken Windows

So, undefine `current_user` before creating the shim, which will keep
the output of the test runs free of this warning.  Hopefully encourage
folks to keep the test run clean and tidy, and stave off [software
entropy][1]

Thanks to @cupakromer for pointing out `undef_method` instead of
juggling `$VERBOSE`.

[1]: https://pragprog.com/the-pragmatic-programmer/extracts/software-entropy