Lockdown 'oauth2' to version under 2
Closes #50

Christopher S. and I were able to track down the error:

> OAuth2::Error - Could not find application:
> #<SnakyHash::StringKeyed error="Could not find application">:

The likely culprit is that the auth call is not including any
identifying data. We didn't find more specific information about this,
but the regression happens between v1.4.9 and >= v2.

We verified that locking down the "oauth2" gem to under 2 fixes the
error. Later on, we'll probably need to address this issue when we want
to use versions 2 and later.